[Refactor] CommonDateTranslator 패키지 변경
